requirements.txt
oca-odoo-pre-commit-hooks
